# backend/tests/fixtures/api_contract.json
{
  "components": {
    "schemas": {
...
      "Body_upload_pgn_api_imports_post": {
        "properties": {
          "analyze": {
            "default": true,
            "title": "Analyze",
            "type": "boolean"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "side": {
            "default": "auto",
            "enum": [
              "auto",
              "white",
              "black"
            ],
            "title": "Side",
            "type": "string"
          },
          "usernames": {
            "default": "",
            "title": "Usernames",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_pgn_api_imports_post",
        "type": "object"
      },
...
      "ChessComRequest": {
        "properties": {
          "analyze": {
            "default": true,
            "title": "Analyze",
            "type": "boolean"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "max_games": {
            "default": 100,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Games",
            "type": "integer"
          },
          "months": {
            "default": 3,
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Months",
            "type": "integer"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "time_class": {
            "default": "rapid",
            "enum": [
              "rapid",
              "blitz",
              "bullet",
              "daily",
              "all"
            ],
            "title": "Time Class",
            "type": "string"
          },
          "username": {
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "ChessComRequest",
        "type": "object"
      },
...
      "ColdPosition": {
        "additionalProperties": false,
        "properties": {
          "exercise_id": {
            "title": "Exercise Id",
            "type": "string"
          },
          "failed": {
            "title": "Failed",
            "type": "boolean"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "last_attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt Id"
          },
          "legal_moves": {
            "items": {
              "$ref": "#/components/schemas/LegalMove"
            },
            "title": "Legal Moves",
            "type": "array"
          },
          "orientation": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Orientation",
            "type": "string"
          },
          "practice_only": {
            "title": "Practice Only",
            "type": "boolean"
          },
          "previous_reviews": {
            "title": "Previous Reviews",
            "type": "integer"
          },
          "review_reason": {
            "enum": [
              "new",
              "resume",
              "learning",
              "relearning",
              "review",
              "practice"
            ],
            "title": "Review Reason",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "last_attempt_id",
          "exercise_id",
          "fen",
          "orientation",
          "failed",
          "review_reason",
          "previous_reviews",
          "practice_only",
          "legal_moves"
        ],
        "title": "ColdPosition",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportError": {
        "additionalProperties": false,
        "properties": {
          "archive": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "game": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game"
          }
        },
        "required": [
          "error"
        ],
        "title": "ImportError",
        "type": "object"
      },
...
      "JobStarted": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStarted",
        "type": "object"
      },
      "JobStatus": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "JobStatus",
        "type": "object"
      },
...
      "MoveExplanation": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "authority": {
            "enum": [
              "stockfish",
              "curated"
            ],
            "title": "Authority",
            "type": "string"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/Frame"
            },
            "title": "Frames",
            "type": "array"
          },
          "move_san": {
            "title": "Move San",
            "type": "string"
          },
          "move_uci": {
            "title": "Move Uci",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "orientation": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Orientation",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Score"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "version": {
            "default": "1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "attempt_id",
          "authority",
          "accepted",
          "move_uci",
          "move_san",
          "summary",
          "notes",
          "frames",
          "orientation"
        ],
        "title": "MoveExplanation",
        "type": "object"
      },
      "MoveRequest": {
        "properties": {
          "from_square": {
            "pattern": "^[a-h][1-8]$",
            "title": "From Square",
            "type": "string"
          },
          "promotion": {
            "anyOf": [
              {
                "enum": [
                  "q",
                  "r",
                  "b",
                  "n"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotion"
          },
          "to_square": {
            "pattern": "^[a-h][1-8]$",
            "title": "To Square",
            "type": "string"
          }
        },
        "required": [
          "from_square",
          "to_square"
        ],
        "title": "MoveRequest",
        "type": "object"
      },
...
      "PgnImportResult": {
        "additionalProperties": false,
        "properties": {
          "duplicates": {
            "title": "Duplicates",
            "type": "integer"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportError"
            },
            "title": "Errors",
            "type": "array"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          }
        },
        "required": [
          "import_id",
          "processed",
          "imported",
          "duplicates",
          "job_id",
          "errors"
        ],
        "title": "PgnImportResult",
        "type": "object"
      },
      "PracticeQueueItem": {
        "additionalProperties": false,
        "properties": {
          "exercise_id": {
            "title": "Exercise Id",
            "type": "string"
          }
        },
        "required": [
          "exercise_id"
        ],
        "title": "PracticeQueueItem",
        "type": "object"
      },
...
      "ReviewQueueItem": {
        "additionalProperties": false,
        "properties": {
          "due": {
            "title": "Due",
            "type": "string"
          },
          "exercise_id": {
            "title": "Exercise Id",
            "type": "string"
          },
          "new": {
            "title": "New",
            "type": "boolean"
          }
        },
        "required": [
          "exercise_id",
          "due",
          "new"
        ],
        "title": "ReviewQueueItem",
        "type": "object"
      },
...
      "Score": {
        "properties": {
          "kind": {
            "enum": [
              "cp",
              "mate"
            ],
            "title": "Kind",
            "type": "string"
          },
          "mate_given": {
            "default": false,
            "title": "Mate Given",
            "type": "boolean"
          },
          "value": {
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "Score",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fieldwork API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/imports": {
      "post": {
        "operationId": "upload_pgn_api_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_pgn_api_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgnImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Pgn"
      }
    },
    "/api/imports/chesscom": {
      "post": {
        "operationId": "import_chesscom_api_imports_chesscom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChessComRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Chesscom"
      }
    },
...
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job"
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job"
      }
    },
    "/api/practice/queue": {
      "get": {
        "operationId": "focused_queue_api_practice_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "skill_id",
            "required": true,
            "schema": {
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PracticeQueueItem"
                  },
                  "title": "Response Focused Queue Api Practice Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Focused Queue"
      }
    },
    "/api/review/queue": {
      "get": {
        "operationId": "review_queue_api_review_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "last_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItem"
                  },
                  "title": "Response Review Queue Api Review Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Queue"
      }
    },
    "/api/review/sessions/{session_id}/explanation": {
      "get": {
        "operationId": "review_explanation_api_review_sessions__session_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "attempt_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attempt Id"
            }
          },
          {
            "in": "query",
            "name": "solution",
            "required": false,
            "schema": {
              "default": false,
              "title": "Solution",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveExplanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Explanation"
      }
    },
    "/api/review/sessions/{session_id}/move": {
      "post": {
        "operationId": "move_api_review_sessions__session_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFeedback"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move"
      }
    },
    "/api/review/sessions/{session_id}/reveal": {
      "post": {
        "operationId": "show_move_api_review_sessions__session_id__reveal_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFeedback"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show Move"
      }
    },
    "/api/review/{exercise_id}/start": {
      "post": {
        "operationId": "begin_review_api_review__exercise_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "exercise_id",
            "required": true,
            "schema": {
              "title": "Exercise Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "focus_skill_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Focus Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColdPosition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Begin Review"
      }
    },
...
  }
}